refactors code adding utility classes, and some minor fixes
