+++ app/components/ReproChecklist/ChecklistItem/ChecklistItem.js
@@ -78,6 +78,8 @@ function ChecklistItem(props) {
     }
 
     setAddAsset(false);
+    setAssetTitle('');
+    setAssetDescription('');
   }
 
   const handleCopy = (urlId, hyperlink) => {
@@ -305,7 +307,7 @@ function ChecklistItem(props) {
                 <div className={`${styles.imageContent} ${showImages ? styles.show : ''}`}>
                   <ul>
                     {item.images.map((image) => (
-                      <li key={image} className={styles.image}>
+                      <li key={image.id} className={styles.image}>
                         <div className={styles.imageHeader}>
                           <span className={styles.imageText}>{image.title}</span>
                           <Delete
@@ -389,7 +391,7 @@ function ChecklistItem(props) {
 
 ChecklistItem.propTypes = {
   item: PropTypes.shape({
-    id: PropTypes.string.isRequired,
+    id: PropTypes.number.isRequired,
     name: PropTypes.string.isRequired,
     statement: PropTypes.string.isRequired,
     answer: PropTypes.bool.isRequired,
